requirements.txt
flask
llama-index
llama-index-llms-ollama
llama-index-embeddings-huggingface
llama-index-vector-stores-qdrant
qdrant-client
sentence-transformers
fpdf2
pymupdf
requests
